Edit Client Profile added

+++ account/models.py
@@ -19,7 +19,7 @@ class Barber(models.Model):
     schedule = models.TextField(null=True)#display schedule in form of text
     description = models.TextField(null=True) #other wanted information
     
-    avgRating = models.IntegerField(null=True)#averaged Rating
+    avgRating = models.FloatField(null=True)#averaged Rating
     profilePic = models.CharField(max_length=200, null=True)#link to prof pic
     def __str__(self):
         return 'Barber Name: %s %s' % (self.firstName, self.lastName)
